refactor(pages): extract duplicated search and heading logic
DocsPage and BlogPage each carried an identical copy of the Cmd+K command
palette (open/close + query state, the keyboard shortcut, focus-on-open,
and result navigation) plus a local extractHeadings function.

Pull the palette into useCommandSearch / useSearchKeyboardNav and the
heading parser into a shared src/utils/extractHeadings util, so both pages
consume one implementation.
